# [Enhancement](runtime filter)optimize for runtime filter