--- a/be/src/exprs/runtime_filter.cpp
+++ b/be/src/exprs/runtime_filter.cpp
@@ -31,12 +31,14 @@
 #include "exprs/literal.h"
 #include "exprs/minmax_predicate.h"
 #include "gen_cpp/internal_service.pb.h"
+#include "runtime/define_primitive_type.h"
 #include "runtime/large_int_value.h"
 #include "runtime/primitive_type.h"
 #include "runtime/runtime_filter_mgr.h"
 #include "runtime/runtime_state.h"
 #include "util/runtime_profile.h"
 #include "util/string_parser.hpp"
+#include "vec/columns/column.h"
 #include "vec/exprs/vbloom_predicate.h"
 #include "vec/exprs/vexpr.h"
 #include "vec/exprs/vruntimefilter_wrapper.h"
@@ -489,8 +491,41 @@ public:
             break;
         }
     }
+
+    void insert_fixed_len(const char* data, const int* offsets, int number) {
+        switch (_filter_type) {
+        case RuntimeFilterType::IN_FILTER: {
+            if (_is_ignored_in_filter) {
+                break;
+            }
+            _hybrid_set->insert_fixed_len(data, offsets, number);
+            break;
+        }
+        case RuntimeFilterType::MINMAX_FILTER: {
+            _minmax_func->insert_fixed_len(data, offsets, number);
+            break;
+        }
+        case RuntimeFilterType::BLOOM_FILTER: {
+            _bloomfilter_func->insert_fixed_len(data, offsets, number);
+            break;
+        }
+        case RuntimeFilterType::IN_OR_BLOOM_FILTER: {
+            if (_is_bloomfilter) {
+                _bloomfilter_func->insert_fixed_len(data, offsets, number);
+            } else {
+                _hybrid_set->insert_fixed_len(data, offsets, number);
+            }
+            break;
+        }
+        default:
+            DCHECK(false);
+            break;
+        }
+    }
+
     void insert(const StringRef& value) {
         switch (_column_return_type) {
+        // todo: rethink logic of hll/bitmap/date
         case TYPE_DATE:
         case TYPE_DATETIME: {
             // DateTime->DateTimeValue
@@ -518,6 +553,16 @@ public:
         default:
             insert(reinterpret_cast<const void*>(value.data));
             break;
+        }
+    }
+
+    void insert_batch(const vectorized::ColumnPtr column, const std::vector<int>& rows) {
+        if (IRuntimeFilter::enable_use_batch(_column_return_type)) {
+            insert_fixed_len(column->get_raw_data().data, rows.data(), rows.size());
+        } else {
+            for (int index : rows) {
+                insert(column->get_data_at(index));
+            }
         }
     }
 
@@ -1000,7 +1045,7 @@ private:
     int32_t _max_in_num = -1;
     std::unique_ptr<MinMaxFuncBase> _minmax_func;
     std::unique_ptr<HybridSetBase> _hybrid_set;
-    std::shared_ptr<IBloomFilterFuncBase> _bloomfilter_func;
+    std::shared_ptr<BloomFilterFuncBase> _bloomfilter_func;
     bool _is_bloomfilter = false;
     bool _is_ignored_in_filter = false;
     std::string* _ignored_in_filter_msg = nullptr;
@@ -1027,6 +1072,12 @@ void IRuntimeFilter::insert(const void* data) {
 void IRuntimeFilter::insert(const StringRef& value) {
     DCHECK(is_producer());
     _wrapper->insert(value);
+}
+
+void IRuntimeFilter::insert_batch(const vectorized::ColumnPtr column,
+                                  const std::vector<int>& rows) {
+    DCHECK(is_producer());
+    _wrapper->insert_batch(column, rows);
 }
 
 Status IRuntimeFilter::publish() {
@@ -1255,7 +1306,7 @@ void IRuntimeFilter::init_profile(RuntimeProfile* parent_profile) {
 }
 
 void IRuntimeFilter::update_runtime_filter_type_to_profile() {
-    if (_profile.get() != nullptr) {
+    if (_profile != nullptr) {
         _profile->add_info_string("RealRuntimeFilterType",
                                   ::doris::to_string(_wrapper->get_real_type()));
     }