--- a/be/src/exprs/runtime_filter.h
+++ b/be/src/exprs/runtime_filter.h
@@ -139,6 +139,7 @@ public:
     // only used for producer
     void insert(const void* data);
     void insert(const StringRef& data);
+    void insert_batch(vectorized::ColumnPtr column, const std::vector<int>& rows);
 
     // publish filter
     // push filter to remote node or push down it to scan_node
@@ -175,7 +176,7 @@ public:
     bool is_producer() const { return _role == RuntimeFilterRole::PRODUCER; }
     bool is_consumer() const { return _role == RuntimeFilterRole::CONSUMER; }
     void set_role(const RuntimeFilterRole role) { _role = role; }
-    int expr_order() { return _expr_order; }
+    int expr_order() const { return _expr_order; }
 
     // only used for consumer
     // if filter is not ready for filter data scan_node
@@ -209,7 +210,7 @@ public:
     void set_ignored() { _is_ignored = true; }
 
     // for ut
-    bool is_ignored() { return _is_ignored; }
+    bool is_ignored() const { return _is_ignored; }
 
     void set_ignored_msg(std::string& msg) { _ignored_msg = msg; }
 
@@ -230,6 +231,10 @@ public:
     void set_push_down_profile();
 
     void ready_for_publish();
+
+    static bool enable_use_batch(PrimitiveType type) {
+        return is_int_or_bool(type) || is_float_or_double(type);
+    }
 
 protected:
     // serialize _wrapper to protobuf
@@ -282,7 +287,7 @@ protected:
 
     // Indicate whether runtime filter expr has been ignored
     bool _is_ignored;
-    std::string _ignored_msg = "";
+    std::string _ignored_msg;
 
     // some runtime filter will generate
     // multiple contexts such as minmax filter