--- a/be/src/util/hash_util.hpp
+++ b/be/src/util/hash_util.hpp
@@ -45,6 +45,14 @@ namespace doris {
 // Utility class to compute hash values.
 class HashUtil {
 public:
+    template <typename T>
+    static uint32_t fixed_len_to_uint32(T value) {
+        if constexpr (sizeof(T) <= sizeof(uint32_t)) {
+            return value;
+        }
+        return std::hash<T>()(value);
+    }
+
     static uint32_t zlib_crc_hash(const void* data, int32_t bytes, uint32_t hash) {
         return crc32(hash, (const unsigned char*)data, bytes);
     }