--- a/be/src/vec/columns/column.h
+++ b/be/src/vec/columns/column.h
@@ -299,6 +299,7 @@ public:
     /// This is for calculating the memory size for vectorized serialization of aggregation keys.
     virtual size_t get_max_row_byte_size() const {
         LOG(FATAL) << "get_max_row_byte_size not supported";
+        return 0;
     }
 
     virtual void serialize_vec(std::vector<StringRef>& keys, size_t num_rows,