--- a/be/src/vec/exec/scan/new_olap_scanner.h
+++ b/be/src/vec/exec/scan/new_olap_scanner.h
@@ -43,7 +43,7 @@ public:
 public:
     Status prepare(const TPaloScanRange& scan_range, const std::vector<OlapScanRange*>& key_ranges,
                    VExprContext** vconjunct_ctx_ptr, const std::vector<TCondition>& filters,
-                   const std::vector<std::pair<string, std::shared_ptr<IBloomFilterFuncBase>>>&
+                   const std::vector<std::pair<string, std::shared_ptr<BloomFilterFuncBase>>>&
                            bloom_filters,
                    const std::vector<FunctionFilter>& function_filters);
 
@@ -58,7 +58,7 @@ private:
 
     Status _init_tablet_reader_params(
             const std::vector<OlapScanRange*>& key_ranges, const std::vector<TCondition>& filters,
-            const std::vector<std::pair<string, std::shared_ptr<IBloomFilterFuncBase>>>&
+            const std::vector<std::pair<string, std::shared_ptr<BloomFilterFuncBase>>>&
                     bloom_filters,
             const std::vector<FunctionFilter>& function_filters);