--- a/be/src/vec/exprs/vbloom_predicate.h
+++ b/be/src/vec/exprs/vbloom_predicate.h
@@ -37,12 +37,12 @@ public:
         return pool->add(new VBloomPredicate(*this));
     }
     const std::string& expr_name() const override;
-    void set_filter(std::shared_ptr<IBloomFilterFuncBase>& filter);
+    void set_filter(std::shared_ptr<BloomFilterFuncBase>& filter);
 
-    std::shared_ptr<IBloomFilterFuncBase> get_bloom_filter_func() const override { return _filter; }
+    std::shared_ptr<BloomFilterFuncBase> get_bloom_filter_func() const override { return _filter; }
 
 private:
-    std::shared_ptr<IBloomFilterFuncBase> _filter;
+    std::shared_ptr<BloomFilterFuncBase> _filter;
     std::string _expr_name;
 };
 } // namespace doris::vectorized