# Proposed changes

1. optimize build runtime filter.
2. change hash function of some numeric type in bloom filter.



TPCH-SF100 | Before | After
-- | -- | --
q1 | 11850.0 | 11756.0
q2 | 604.8 | 430.2
q3 | 3726.4 | 3195.6
q4 | 1569.8 | 1185.0
q5 | 2875.6 | 2492.8
q6 | 971.6 | 982.8
q7 | 1439.0 | 1281.4
q8 | 1481.8 | 1085.2
q9 | 20764.0 | 20939.0
q10 | 2113.2 | 1895.8
q11 | 615.2 | 617.8
q12 | 887.6 | 749.6
q13 | 3658.2 | 3562.4
q14 | 705.4 | 653.6
q15 | 612.8 | 605.8
q16 | 910.2 | 878.6
q17 | 4969.4 | 1527.2
q18 | 4027.2 | 3957.6
q19 | 429.4 | 445.8
q20 | 1685.0 | 1574.6
q21 | 4800.4 | 4551.6
q22 | 1261.4 | 1134.8
SUM | 71958.4 | 65503.2




## Problem summary

Describe your changes.

## Checklist(Required)

1. Does it affect the original behavior: 
    - [ ] Yes
    - [ ] No
    - [ ] I don't know
3. Has unit tests been added:
    - [ ] Yes
    - [ ] No
    - [ ] No Need
4. Has document been added or modified:
    - [ ] Yes
    - [ ] No
    - [ ] No Need
5. Does it need to update dependencies:
    - [ ] Yes
    - [ ] No
6. Are there any changes that cannot be rolled back:
    - [ ] Yes (If Yes, please explain WHY)
    - [ ] No

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...